requirements.txt
flask
pandas
scikit-learn
tk
mysql-connector-python